Separate cxx_atomic storage component from atomic_base, create atomic_nvrtc header
