+++ libcxx/include/support/atomic/atomic_base.h
@@ -11,89 +11,7 @@
 #ifndef _LIBCUDACXX_ATOMIC_BASE_H
 #define _LIBCUDACXX_ATOMIC_BASE_H
 
-template <typename _Tp, int _Sco>
-struct __cxx_atomic_base_impl {
-  using __underlying_t = _Tp;
-  static constexpr int __sco = _Sco;
-
-  _LIBCUDACXX_CONSTEXPR
-  __cxx_atomic_base_impl() _NOEXCEPT = default;
-
-  _LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR explicit
-  __cxx_atomic_base_impl(_Tp value) _NOEXCEPT : __a_value(value) {}
-
-  _ALIGNAS(sizeof(_Tp)) _Tp __a_value;
-};
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-_Tp* __cxx_get_underlying_atomic(__cxx_atomic_base_impl<_Tp, _Sco> * __a) _NOEXCEPT {
-  return &__a->__a_value;
-}
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-volatile _Tp* __cxx_get_underlying_atomic(__cxx_atomic_base_impl<_Tp, _Sco> volatile* __a) _NOEXCEPT {
-  return &__a->__a_value;
-}
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-const _Tp* __cxx_get_underlying_atomic(__cxx_atomic_base_impl<_Tp, _Sco> const* __a) _NOEXCEPT {
-  return &__a->__a_value;
-}
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-const volatile _Tp* __cxx_get_underlying_atomic(__cxx_atomic_base_impl<_Tp, _Sco> const volatile* __a) _NOEXCEPT {
-  return &__a->__a_value;
-}
-
-template <typename _Tp, int _Sco>
-struct __cxx_atomic_ref_base_impl {
-  using __underlying_t = _Tp;
-  static constexpr int __sco = _Sco;
-
-  _LIBCUDACXX_CONSTEXPR
-  __cxx_atomic_ref_base_impl() _NOEXCEPT = default;
-
-  _LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR explicit
-  __cxx_atomic_ref_base_impl(_Tp value) _NOEXCEPT : __a_value(value) {}
-
-  _Tp* __a_value;
-};
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-_Tp* __cxx_get_underlying_atomic(__cxx_atomic_ref_base_impl<_Tp, _Sco>* __a) _NOEXCEPT {
-  return __a->__a_value;
-}
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-volatile _Tp* __cxx_get_underlying_atomic(__cxx_atomic_ref_base_impl<_Tp, _Sco> volatile* __a) _NOEXCEPT {
-  return __a->__a_value;
-}
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-const _Tp* __cxx_get_underlying_atomic(__cxx_atomic_ref_base_impl<_Tp, _Sco> const* __a) _NOEXCEPT {
-  return __a->__a_value;
-}
-
-template <typename _Tp, int _Sco>
-_LIBCUDACXX_INLINE_VISIBILITY _LIBCUDACXX_CONSTEXPR
-const volatile _Tp* __cxx_get_underlying_atomic(__cxx_atomic_ref_base_impl<_Tp, _Sco> const volatile* __a) _NOEXCEPT {
-  return __a->__a_value;
-}
-
-template <typename _Tp>
-_LIBCUDACXX_INLINE_VISIBILITY auto __cxx_atomic_base_unwrap(_Tp* __a) _NOEXCEPT -> decltype(__cxx_get_underlying_atomic(__a)) {
-  return __cxx_get_underlying_atomic(__a);
-}
-
-template <typename _Tp>
-using __cxx_atomic_underlying_t = typename _Tp::__underlying_t;
+#include "cxx_atomic.h"
 
 _LIBCUDACXX_INLINE_VISIBILITY inline _LIBCUDACXX_CONSTEXPR int __cxx_atomic_order_to_int(memory_order __order) {
   // Avoid switch statement to make this a constexpr.
