+++ libcxx/include/support/atomic/atomic_nvrtc.h
@@ -1,0 +1,16 @@
+// -*- C++ -*-
+//===----------------------------------------------------------------------===//
+//
+// Part of libcu++, the C++ Standard Library for your entire system,
+// under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef _LIBCUDACXX_ATOMIC_NVRTC_H
+#define _LIBCUDACXX_ATOMIC_NVRTC_H
+
+#include "cxx_atomic.h"
+
+#endif // _LIBCUDACXX_ATOMIC_NVRTC_H
